Fix bug of Number too big

+++ Skyblock Bazzar Tracker/Products_api_info.cs
@@ -6,20 +6,20 @@
         {
             public string productId { get; set; }
             public double sellPrice { get; set; }
-            public int sellVolume { get; set; }
-            public int sellMovingWeek { get; set; }
-            public int sellOrders { get; set; }
+            public Int64 sellVolume { get; set; }
+            public Int64 sellMovingWeek { get; set; }
+            public Int64 sellOrders { get; set; }
             public double buyPrice { get; set; }
-            public int buyVolume { get; set; }
-            public int buyMovingWeek { get; set; }
-            public int buyOrders { get; set; }
+            public Int64 buyVolume { get; set; }
+            public Int64 buyMovingWeek { get; set; }
+            public Int64 buyOrders { get; set; }
         }
 
         public class Sell_summery_class
         {
-            public int amount { get; set; }
+            public Int64 amount { get; set; }
             public double pricePerUnit { get; set; }
-            public int orders { get; set; }
+            public Int64 orders { get; set; }
         }
 
         public Sell_summery_class[] sell_summary { get; set; }
